Refactor playRound() to return numerical results
Change the playRound function in the Rock, Paper, Scissors game to
return numerical results for more straightforward scoring.

The playRound function was returning string messages, making
difficult to keep score in the game function. Now, the playRound
function returns -1 for a computer win, 1 for a player win, and 0 for
a tie. User feedback is still provided through console.log messages
during each round.This refactor prepares us for score-keeping in the
game function.